+++ tests/core/routes/conftest.py
@@ -2,10 +2,6 @@ import base64
 import json
 
 import pytest
-from sqlalchemy.orm import Session
-
-from papermerge.core import db, schemas
-from papermerge.core.models import User
 
 
 def b64e(s):
@@ -24,13 +20,3 @@ def token():
     payload = b64e(json_str)
 
     return f"ignore_me.{payload}.ignore_me_too"
-
-
-@pytest.fixture
-def custom_field_cf1(db_session: Session, user: User):
-    return db.create_custom_field(
-        db_session,
-        name="cf1",
-        type=schemas.CustomFieldType.text,
-        user_id=user.id,
-    )
